Move notification module
