+++ src/rust/ide/src/controller.rs
@@ -17,7 +17,6 @@
 
 pub mod graph;
 pub mod module;
-pub mod notification;
 pub mod project;
 pub mod text;
